# services/swagger/users.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "openapi document",
        "version": "0.0.1"
    },
    "servers": [
        {
            "url": "http://localhost:8081",
            "description": "local api server"
        }
    ],
    "components": {
        "schemas": {
            "apiResponse": {
                "type": "object",
                "required": [
                    "success",
                    "message"
                ],
                "properties": {
                    "success": {
                        "type": "boolean",
                        "readOnly": true
                    },
                    "message": {
                        "type": "string",
                        "readOnly": true
                    },
                    "data": {
                        "type": "object",
                        "readOnly": true
                    }
                }
            },
            "userAddRequest": {
                "type": "object",
                "required": [
                    "id",
                    "email",
                    "given_name",
                    "family_name",
                    "prefix",
                    "suffix"
                ],
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid",
                        "readOnly": true
                    },
                    "email": {
                        "type": "string",
                        "format":"email",
                        "readOnly": true
                    },
                    "given_name": {
                        "type": "string",
                        "readOnly": true
                    },
                    "family_name": {
                        "type": "string",
                        "readOnly": true
                    },
                    "prefix": {
                        "type":"string",
                        "readOnly":true
                    },
                    "suffix": {
                        "type":"string",
                        "readOnly":true
                    }
                },
                "example": {
                    "id": "47d7164a-fc10-4216-96c5-e2ad34ad66a4",
                    "email": "test@test.com",
                    "given_name": "test_given_name",
                    "family_name": "test_family_name",
                    "prefix": "",
                    "suffix": ""
                }
            },
            "userSetActiveRequest": {
                "type":"object",
                "required":[
                    "id",
                    "active"
                ],
                "properties":{
                    "id":{
                        "type":"string",
                        "format":"uuid",
                        "readOnly":true
                    },
                    "active":{
                        "type":"boolean",
                        "readOnly":true
                    }
                },
                "example":{
                    "id":"47d7164a-fc10-4216-96c5-e2ad34ad66a4",
                    "active":true
                }
            },
            "userSetPasswordRequest": {
                "type":"object",
                "required":[
                    "id",
                    "password"
                ],
                "properties":{
                    "id":{
                        "type":"string",
                        "format":"uuid",
                        "readOnly":true
                    },
                    "password": {
                        "type":"boolean",
                        "readOnly":true
                    }
                },
                "example":{
                    "id":"47d7164a-fc10-4216-96c5-e2ad34ad66a4",
                    "password":"ThisIsaPassword8*"
                }
            },
            "userAuthenticateRequest": {
                "type":"object",
                "required":[
                    "email",
                    "password"
                ],
                "properties":{
                    "email":{
                        "type":"string",
                        "format":"email",
                        "readOnly":true
                    },
                    "password":{
                        "type":"string",
                        "readOnly":true
                    }
                },
                "example":{
                    "email":"test@test.com",
                    "password":"ThisIsaPassword8*9"
                }
            }
        },
        "securitySchemes": {
            "bearer": {
                "type": "http",
                "description": "Bearer authentication",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer"
            }
        }
    },
    "paths": {
        "/user/add": {
            "summary": "add a user",
            "post": {
                "summary": "POST method add a user",
                "tags": ["user"],
                "operationId": "user_add",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/userAddRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "domain added",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/apiResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/user/set/active": {
            "summary": "set user active status",
            "post": {
                "summary": "POST method to set user active status",
                "tags": ["user"],
                "operationId": "user_set_active",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/userSetActiveRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "user active status set",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/apiResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/user/set/password": {
            "summary": "set user password",
            "post": {
                "summary": "POST method to set user password",
                "tags": ["user"],
                "operationId": "user_set_password",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/userSetPasswordRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "user set password",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/apiResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/user/authenticate": {
            "summary": "authenticate user",
            "post": {
                "summary": "POST method authenticate user",
                "tags": ["user"],
                "operationId": "user_authenticate",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/userAuthenticateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "user authenticate",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/apiResponse"
                                }
                            }
                        }
                    }
                }
            }
        }
    }
}